Support measures with different data types (enhancement #3) (#22)
Implements #3 and foundation for #5 

Currently the measures store is a bit of a mess and supports only double
values (stored as IQuantity, with their unit of measure). We need to
support other types, at least `string`. We also need to support
different storage strategies (in-memory as it is now but also local DB
or Cloud, see #5).

This PR implements #3 and lay the foundation to implement #5 (through
the use of `IStorageStrategy` as interface for the underlying
implementation).

For now we added a `MeasureValue` struct that hides the effective value,
we begin with `IQuantity` for floating point and `string` but this
should make it easier to add more types in the future.

Because this is a foundational change we'll also need to change the
`tinkwell.store` protobuf contract and ALL the other services using this
contract. Because of this, I used this chance to change the response
types when searching for measures: instead of returning the whole DB
(potentially big) we implemented `stream` which in C# is an asynchronous
enumerable: items can be consumed one by one without loading everything
in memory.