CLI utility to create configuration files from templates (feature #6) (#23)
Implements #6 

Added command `tw templates`. It supports: `create` (to create a new
template), `inspect` (to show all the properties of a given template),
`list` (to list all the available templates) and `use` (to use a
template to generate _something_).

Templates are Liquid files copied (or appended) to the output folder
with optional variables filled in by the user when `use` applying the
template. It supports saving the given answers to create unattended
generation (see CLI documentation for `tw templates` for details and
examples).

Templates are local only, for now we do not need to support remote
templates!