Unit tests for Reducer (issue #4) (#29)
Linked to #7 , it's needed for #14 

At first my plan was to test Reducer, Reactor and Executor only with
integration tests. My rationale was that they're mostly plumbing around
other services and the little they add isn't worthy the effort (at least
for now), considering that they're going to be tested anyway with
integration tests.

However, to properly support an isolated environment (usable in CI) for
integration tests we need some refactoring around service discovery and
I do not want to make any refactoring without at least some unit testing
in place so...here it is.