Integration tests (enhancement #14)  (#32)
Implements #14 

Dropped the idea of integration tests in C#. The new test runner is a
small Python script (in `Tests/Integration`. How it works:

* It collects all the tests in `./tests`. Test files are Python files
starting with `test_`. If they export a `TRAIT` attribute then they can
be filtered by trait (for example to run fewer tests locally and the
whole suite in CI or to run integration vs smoke tests).
* It runs each test sequentially, spinning up a new isolated environment
for each one.
* For each test it setups an isolated environment where the application
can run without using any existing data in the system.
* It generates a self-signed certificate for the application to use.
* If there is a directory with the same name of the test then its
content is copied as-is in the isolated environment.
* The test do whatever it has to do (it could do gRPC calls but it's
probably easier to use `tw` as an _interface_ with the application and
it returns failed/pass.
    * The test runner deletes any temporary file.
* The runner displays the result. Its exit code is 1 if any test failed.

It requires:

* Python 3
* .NET 9

It's a simple command line application with these options:

* `--app-path <path>`: path to the directory where the application is.
Required.
* `--test-dir<path>`: path to the directory where the test files are
(default `./tests`).
* `--trait<trait>`: if specified then only tests with this trait are
executed.
* `--test-name<path>`: if specified then only the test with the
specified name is executed.
* `--keep-temp-dir`: if specified then the temporary folder(s) are not
deleted at the end of the test.
* `--verbose`: shows more details about what's going on (including the
full log from the Supervisor).